hide the ugliness of osd vertical overlap
there seems to be no helping the fact that the bottom two lines are
painted with the top buffer. i'm guessing this is because the interrupt
happens at the wrong time, but i'm not sure.

because i can't fix it, just reduce the screen usage at the top and
bottom so that it appears normal